requirements.txt
streamlit
langchain
langchain-google-genai
chromadb
pypdf
python-dotenv
google-generativeai
